Introduce GMD + ATD (#479)
Introduce Gradle Managed Device definitions to modules that have instrumented tests.

At the moment, defining it as a convention plugin isn't possible due to the bug. That bug is now fixed and will be included in Android Studio Flamingo.
Once Flamingo becomes stable, convert the GMD definitions as a convention plugin (#523)

Change-Id: I0866369e3d0bbe148ca5ec1f92bad59239a347b8