Android-CI workflow -> GMD refactor (#537)
This PR is about updating the operating system to macOS-12 as the previous one was going to be deprecated, as well as updating the checkout to v3 to avoid the warnings in the panel; adding setup SDK action to accept the license which was missing from the beginning.

I also added imports instead of adding them directly with the codes.

As a reference, there was another PR with a similar topic, but it wasn't complete: #483